[이동] 예제 디렉토리 상위폴더로 이동
