+++ src/main/java/jpabook/exercise/DeliveryStatus.java
@@ -1,4 +1,4 @@
-package jpabook.mapping.exercise;
+package jpabook.exercise;
 
 public enum DeliveryStatus {
     READY, COMP
