+++ src/main/java/jpabook/exercise/Member.java
@@ -1,29 +1,28 @@
-package jpabook.mapping.exercise;
+package jpabook.exercise;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import javax.persistence.OneToOne;
+import javax.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
-public class Delivery {
+public class Member extends BaseEntity {
     @Id
     @GeneratedValue
-    @Column(name = "DELIVERY_ID")
+    @Column(name = "MEMBER_ID")
     private Long id;
 
-    @OneToOne(mappedBy = "DELIVERY_ID")
-    private Order order;
+    private String name;
 
     private String city;
     private String street;
     private String zipcode;
 
-    @Enumerated(EnumType.STRING)
-    private DeliveryStatus status;
+    @OneToMany(mappedBy = "member")
+    private List<Order> orders = new ArrayList<Order>();
 
     public Long getId() {
         return id;
@@ -33,12 +32,12 @@ public class Delivery {
         this.id = id;
     }
 
-    public Order getOrder() {
-        return order;
+    public String getName() {
+        return name;
     }
 
-    public void setOrder(Order order) {
-        this.order = order;
+    public void setName(String name) {
+        this.name = name;
     }
 
     public String getCity() {
@@ -65,11 +64,11 @@ public class Delivery {
         this.zipcode = zipcode;
     }
 
-    public DeliveryStatus getStatus() {
-        return status;
+    public List<Order> getOrders() {
+        return orders;
     }
 
-    public void setStatus(DeliveryStatus status) {
-        this.status = status;
+    public void setOrders(List<Order> orders) {
+        this.orders = orders;
     }
 }
